requirements.txt
flask==3.0.3
gunicorn==23.0.0
opencv-python-headless==4.10.0.84
numpy==2.0.2
pillow==10.4.0
tqdm==4.66.4
requests==2.32.3
gspread==6.1.4
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1


